Fixes #52
